Put prisma in a central location
Prisma has been given its own 'module'. To use it, import it in your
module with:

    import { prisma } from '../prisma'

Then you can make any calls with the prisma client.

Some other refactoring for authentication and session data is included
in this commit. The session now holds a prisma User object, which can
be easily passed on to any resolvers without having to make more database
lookups.

It should be noted that whenever a database update to the user who is
logged in is made, the user must log out for changes to take effect.